requirements.txt
amqp==2.6.1
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
boto3==1.16.41
botocore==1.19.41
cairocffi==1.2.0
CairoSVG==2.5.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.4
django-csp==3.7
django-otp==1.0.2
django-storages==1.11
django-weasyprint==1.0.2
flower==0.9.7
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
humanize==3.2.0
idna==2.10
jmespath==0.10.0
kombu==4.6.11
mutagen==1.45.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
pydub==0.24.1
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
qrcode==6.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.3
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
tinycss2==1.1.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
